merge: pull request #46
* fix: resolve stations dto and updating

* fix: unit tests for stations

* fix: station mapper

* feat: add alarms REST API endpoints and controllers for managing alarms

* chore: remove pull request trigger from CD workflow for main branch

* fix: restore ListAlarmsController import and rename ListAlarmController to ListAlarmsController

* refactor: update station ID and pagination in telemetry API;

+++ src/infra/database/prisma/repositories/prisma-stations-repository.ts
@@ -41,20 +41,26 @@ export class PrismaStationsRepository
       station,
       parametersIds,
     )
-    await this.prisma.$transaction([
-      this.prisma.stationParameter.deleteMany({
-        where: { stationId: station.id.value },
-      }),
-      this.prisma.station.update({
-        where: { id: station.id.value },
-        data: {
-          ...stationData,
-          ...(stationParameter && {
+
+    if (stationParameter) {
+      await this.prisma.$transaction([
+        this.prisma.stationParameter.deleteMany({
+          where: { stationId: station.id.value },
+        }),
+        this.prisma.station.update({
+          where: { id: station.id.value },
+          data: {
+            ...stationData,
             stationParameter: { create: stationParameter.create },
-          }),
-        },
-      }),
-    ])
+          },
+        }),
+      ])
+    }
+
+    await this.prisma.station.update({
+      where: { id: station.id.value },
+      data: stationData,
+    })
   }
 
   async findMany({
